omg this is annoying

+++ train.py
@@ -36,7 +36,7 @@ def train_loop_adversarial(
         model.eval()
         for param in model.parameters():
             param.requires_grad = False
-        examples = adversary.perturb(examples, labels.type(torch.LongTensor))
+        examples = adversary.perturb(examples, labels)
         for param in model.parameters():
             param.requires_grad = True
         model.train()
